Add files via upload
Hello everyone here I design restaurant web app which is easy to use and navigate. The customer should be able to easily find the necessary information and make their selections. In this web app I mentioned restaurant menu which should be categorized, so that customers can easily find the dishes they are looking for.